InspiredFoodPlatform/requirements.txt
grpcio
googleapis-common-protos
flask
redis
grpcio-tools
grpcio-reflection
fastapi
uvicorn
requests